requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.22.1
biosppy==1.0.0
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.5
cvxopt==1.3.0.1
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
easydev==0.12.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.0
fonttools==4.37.4
fqdn==1.5.1
future==0.18.3
h5py==3.8.0
idna==3.4
imageio==2.25.0
inflection==0.5.1
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
Js2Py==0.74
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nolds==0.5.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
opencv-python==4.7.0.68
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
PyExecJS==1.5.1
Pygments==2.14.0
pyjsparser==2.7.1
PyJWT==2.8.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-http-client==3.3.7
python-json-logger==2.0.4
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.2
secure-smtplib==0.1.1
Send2Trash==1.8.0
sendgrid==6.10.0
shortuuid==1.0.11
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
starkbank-ecdsa==2.2.0
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.1.23.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.8.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.22.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
wfdb==4.1.1
widgetsnbextension==4.0.5
